Rename gaia_catalog_internal to ddl_execution and gaia_catalog_internal_api to catalog_internal
