+++ production/catalog/src/ddl_executor.cpp
@@ -2,6 +2,7 @@
 // Copyright (c) Gaia Platform LLC
 // All rights reserved.
 /////////////////////////////////////////////
+
 #include "ddl_executor.hpp"
 
 #include <memory>
