+++ production/catalog/src/gaia_generate.cpp
@@ -2,6 +2,7 @@
 //// Copyright (c) Gaia Platform LLC
 //// All rights reserved.
 ///////////////////////////////////////////////
+
 #include <memory>
 #include <set>
 #include <vector>
