+++ production/system/tests/test_gaia_system.cpp
@@ -15,10 +15,10 @@
 
 #include "db_test_base.hpp"
 #include "db_test_helpers.hpp"
+#include "ddl_execution.hpp"
 #include "gaia_addr_book.h"
 #include "gaia_catalog.h"
 #include "gaia_catalog.hpp"
-#include "gaia_catalog_internal.hpp"
 #include "gaia_system.hpp"
 #include "rules.hpp"
 #include "triggers.hpp"
