+++ production/tools/gaia_dump/tests/test_gaia_dump.cpp
@@ -5,25 +5,29 @@
 
 #include <fstream>
 #include <iostream>
+
 #include "gtest/gtest.h"
+
+#include "db_test_base.hpp"
+#include "ddl_execution.hpp"
 #include "gaia_airport.h"
 #include "gaia_catalog.hpp"
-#include "gaia_catalog_internal.hpp"
-#include "gaia_parser.hpp"
-#include "db_test_base.hpp"
 #include "gaia_dump.hpp"
+#include "gaia_parser.hpp"
 
 using namespace gaia::catalog;
 using namespace gaia::db;
 using namespace std;
 
-class gaia_dump_test : public db_test_base_t {
+class gaia_dump_test : public db_test_base_t
+{
 };
 
 constexpr gaia_id_t c_start = 1;
 constexpr gaia_id_t c_end = 2;
 
-TEST_F(gaia_dump_test, dump) {
+TEST_F(gaia_dump_test, dump)
+{
     create_database("airport_test");
     ddl::field_def_list_t fields;
     fields.emplace_back(make_unique<ddl::field_definition_t>("name", data_type_t::e_string, 1));
